Split the application officially into two separate applications.
We have finally created two applications, currently with hard-coded ports and host names. Surprisingly, it actually works! We need to have a "connect" button to make one game the leader and the other game the follower (at the beginning at least.)
Furthermore, this initial connection should also communicate anything that would be required to sync up games; names, colors, etc.
While there  is much work to do, this is a huge step forward. The path forward is now clear.