Complete Base Game; Setup for fine-tuning.
We've completed the base game (with networking), although "completed" assumes that all the logic makes sense. This is not the case. The controller and the ui both keep separate solution boards, and preliminary testing has shown that there is the possibility that this solution board is not as complete as previously thought. Additionally, many of the variable names do not make sense - cPort instead of clientPort, etc.
Right now, however, the biggest problem by far is the need for two, separate application starters - Sudoku Primary and Sudoku Secondary. As of this commit, the two are (nearly) identical - the only thing different is the names (Alex Rao and Raeedah Choudhury, respectively) and the colors (RED and BLUE, respectively). In the future, we'd like to be able to specify names, ports, etc. This will require making a unified application structure, and would possibly mean a marginally different application architecture. This will also be the focus of the next commit.